zsh: Rename .zsh* to non-dot names and symlink to dot-names
It is tremendously irksome that zsh doesn't support non-dot-prefixed
names for its files when they're stored in an xdg-compliant directory.
I want to be able to see the files when running ls on the directory.
Accordingly, I have made the required dot-names symlinks to more
descriptive names that actually show up in a directory list.